% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_annotate.R
\name{cluster_annotate}
\alias{cluster_annotate}
\title{annotate the celltype for cluster in the seurat.obj}
\usage{
cluster_annotate(
  seurat_object,
  cluster_markers = NULL,
  filename = "",
  width = 4,
  height = 2.5,
  n_col = NULL
)
}
\arguments{
\item{seurat_object}{annotation of the seurat.obj}

\item{cluster_markers}{marker genes of celltypes}

\item{filename}{the name of the generated file}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{n_col}{ncol for the vlnplot}
}
\value{
a seurat.obj with annotation
}
\description{
annotate the celltype for cluster in the seurat.obj
}
\examples{
\dontrun{
data(common_cluster,package="LIANLAB")
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
pbmc_1k <- cluster_annotate( seurat_object = pbmc_1k, cluster_markers = common_cluster)
}
}
