% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_heatmap.R
\name{Get_ave_heatmap}
\alias{Get_ave_heatmap}
\title{To draw the average expression of the cluster}
\usage{
Get_ave_heatmap(seurat.obj, genes, split = c(T, F), colorm = NULL)
}
\arguments{
\item{seurat.obj}{the seurat object}

\item{genes}{the genes you interested}

\item{split}{whether to split the heatmap}

\item{colorm}{the data.frame including clusters(first col) and colors(second col)}
}
\description{
To draw the average expression of the cluster
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
genes = c('CD8A','CD3','CD4')
Get_ave_heatmap(pbmc_1k,genes,split=T)
}
}
