% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_BASIC.R
\name{Get_BASIC_drawing}
\alias{Get_BASIC_drawing}
\title{To get basic drawing, including t-SNE, split figure and featureplot}
\usage{
Get_BASIC_drawing(
  object,
  filename = NULL,
  genenames,
  pt.size = 0.6,
  split_by = "samples",
  width = 7,
  height = 6,
  ncol = 2
)
}
\arguments{
\item{object}{seurat.obj}

\item{filename}{the name of the generated file}

\item{genenames}{the genes you interested}

\item{pt.size}{size of the point}

\item{split_by}{split the figure by}

\item{width}{the width of the figure}

\item{height}{the height of the figure}

\item{ncol}{the col of the figure}
}
\description{
To get basic drawing, including t-SNE, split figure and featureplot
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_small.RDS',package = 'LIANLAB')
load(input.file)
Get_BASIC_drawing(pbmc_small,genenames=c('CD8A','CD4'))
}
}
