% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Cor.R
\name{Get_Cor}
\alias{Get_Cor}
\title{Compare the correlations between clusters in one/two seurat.obj, only in mouse or human}
\usage{
Get_Cor(
  seurat.1,
  seurat.2 = NULL,
  species.1 = "mouse",
  gene = NULL,
  species.2 = "mouse",
  method = "pearson"
)
}
\arguments{
\item{seurat.1}{seurat.obj1}

\item{seurat.2}{seurat.obj2, could be NULL}

\item{species.1}{the species of the seuart.obj1, 'mouse' (default)}

\item{gene}{To Compare the correlations in the genes you interested}

\item{species.2}{the species of the seuart.obj2 (default)}

\item{method}{One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}
}
\value{
the data.frame of the correlations between cluster in two seurat.obj
}
\description{
Compare the correlations between clusters in one/two seurat.obj, only in mouse or human
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
Get_Cor(pbmc_1k,species.1 = 'mouse')
}
}
