% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Seurat.R
\name{Get_for_integrated}
\alias{Get_for_integrated}
\title{To get seurat.obj for integrated}
\usage{
Get_for_integrated(
  object,
  minfeature = 200,
  mincell = 10,
  species = c("h", "m"),
  minFeatureRNA = NULL,
  maxFeatureRNA = NULL,
  is.tumor = F,
  Percent.mt = NULL,
  projectname = NULL
)
}
\arguments{
\item{object}{gene expression matrix}

\item{minfeature}{Include genes where at least this many features are detected.}

\item{mincell}{Include cells where at least this many features are detected.}

\item{species}{the species of the matrix}

\item{minFeatureRNA}{the min cut off the genes count}

\item{maxFeatureRNA}{the max cut off the genes count}

\item{is.tumor}{whether the sample is tumor}

\item{Percent.mt}{the max cut off of the percetnage of the mitochondrial gene}

\item{projectname}{the name of the project}
}
\value{
the seurat.obj
}
\description{
To get seurat.obj for integrated
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_small.csv',package = 'LIANLABDATA')
pbmc_small <- read.csv(input.file, header = T, row.names = 1)

seurat.obj = CreateSeuratObject(counts = pbmc_small, min.features = 200,min.cells = 10)
dim(Abseq.seurat)
See_mRNA(Abseq.seurat)
data[["percent.mt"]] <- PercentageFeatureSet(Abseq.seurat, pattern = "^MT.")  # MT(human)/mt(mouse)
VlnPlot(Abseq.seurat, features = c("nFeature_RNA", "nCount_RNA", "percent.mt"), ncol = 3)
table(Abseq.seurat[["percent.mt"]]<10)

seurat.obj = Get_for_integrated(pbmc_small, minfeature = 200, mincell = 10, species = 'h')
}

}
