% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_GK.R
\name{Get_GK1}
\alias{Get_GK1}
\title{Batch/Specify cell population for GO and KEGG analyse}
\usage{
Get_GK1(
  marker,
  spe_cluster = NULL,
  species = c("m", "h"),
  logfc.cutoff = 0.1,
  p.cutoff = 0.01
)
}
\arguments{
\item{marker}{result of FindAllMarkers(), including cols named avg_logFC, p_val_adj, gene}

\item{spe_cluster}{the clusters you interested}

\item{species}{m/h}

\item{logfc.cutoff}{cut off the logFC}

\item{p.cutoff}{cut off the p value}
}
\description{
Batch/Specify cell population for GO and KEGG analyse
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLABDATA')
markers <- read.csv(input.file,header = T, row.names = 1)
Get_GK1( markers_top,species = 'm')
Get_GK1( markers_top,spe_cluster = 'Naive CD4 T',species = 'm')
}
}
