% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_GK.R
\name{Get_GK2}
\alias{Get_GK2}
\title{GO and KEGG analyse of the result between two clusters}
\usage{
Get_GK2(marker, species = c("m", "h"), logfc.cutoff = 0.1, p.cutoff = 0.01)
}
\arguments{
\item{marker}{result of FindMarkers() or other,including cols named avg_log2FC, p_val_adj, gene}

\item{species}{m/h}

\item{logfc.cutoff}{cut off the logFC}

\item{p.cutoff}{cut off the p value}
}
\description{
GO and KEGG analyse of the result between two clusters
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLABDATA')
markers <- read.csv(input.file,header = T, row.names = 1)
Get_GK2( markers_top,species = 'm')
}
}
