% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_GK.R
\name{Get_GK_genes}
\alias{Get_GK_genes}
\title{Integrated GO/KEGG pathways gene}
\usage{
Get_GK_genes(DATA, keyword = NULL)
}
\arguments{
\item{DATA}{the GO/KEGG result, including cols named geneID and Description}

\item{keyword}{the keywood of the pathways you interested}
}
\value{
the genes in the pathways you choose
}
\description{
Integrated GO/KEGG pathways gene
}
\examples{
\dontrun{
input.file <- system.file('extdata','GO_enrich.csv',package = 'LIANLAB')
markers <- read.csv(input.file,header = T, row.names = 1)
Get_GK_genes(markers)
Get_GK_genes(markers, keyword = c('DNA'))
}
}
