% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{Get_HM_gene_change}
\alias{Get_HM_gene_change}
\title{To change the genes from human/mouse to mouse/human}
\usage{
Get_HM_gene_change(Objects, species = c("h", "m"))
}
\arguments{
\item{Objects}{the matrix to change gene}

\item{species}{species of the matrix genes}
}
\value{
The genes after changed
}
\description{
To change the genes from human/mouse to mouse/human
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_small.csv',package = 'LIANLABDATA')
matrix.obj <- read.csv(input.file,header = T, row.names = 1)
matrix.obj <- Get_HM_gene_change( matrix.obj, species = 'h')
character.obj <- c('Cd4','Cd8a','Cd8b1','Cd3e')
character.obj <- Get_HM_gene_change( character.obj, species = 'm')
}
}
