% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Seurat.R
\name{Get_integrated}
\alias{Get_integrated}
\title{To integrate different batch seurat.obj together}
\usage{
Get_integrated(umi.list)
}
\arguments{
\item{umi.list}{the seurat list}
}
\value{
the integrated seurat.obj
}
\description{
To integrate different batch seurat.obj together
}
\examples{
\dontrun{
seurat.int = Get_integrated(list(seurat.obj1,seurat.obj2,seurat.obj3))
}

}
