% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Monocle.R
\name{Get_monocle}
\alias{Get_monocle}
\title{To analyse the seurat.obj easily}
\usage{
Get_monocle(
  seurat_object,
  gene_ann = NULL,
  n_cell = NULL,
  num_expressed = 1,
  batch_if = NULL,
  remove_batch = F,
  dispersion_empiricals = 0.1,
  mean_expressions = 0.01,
  ordering_genes = NULL
)
}
\arguments{
\item{seurat_object}{seurat.obj}

\item{gene_ann}{gene information rownames are genenames, and a col named 'gene_short_name'}

\item{n_cell}{barcode of the mixed cells, default NULL}

\item{num_expressed}{the lowest gene expression, default}

\item{batch_if}{batch factor, default NULL}

\item{remove_batch}{whether to remove the batch effect caused by the batch factor}

\item{dispersion_empiricals}{dispersion empiricals}

\item{mean_expressions}{mean expressions}

\item{ordering_genes}{marks genes that will be used for clustering in subsequent calls to clusterCells}
}
\value{
a list, including cds and cluster DE genes
}
\description{
To analyse the seurat.obj easily
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
cds.list <- Get_monocle(pbmc_1k)
}
}
