% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Percentplot.R
\name{Get_percentplot}
\alias{Get_percentplot}
\title{Get percentage of seurat.obj}
\usage{
Get_percentplot(seurat_object, filename, colorm = NULL)
}
\arguments{
\item{seurat_object}{seurat.obj, including cols could split}

\item{filename}{the name of the generated file}

\item{colorm}{the colors of the clusters}
}
\description{
Get percentage of seurat.obj
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
Get_percentplot(pbmc_1k)
}
}
