% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{Get_samples_to_group}
\alias{Get_samples_to_group}
\title{To split the cells to 'group' by 'samples'}
\usage{
Get_samples_to_group(seurat.obj, group.list)
}
\arguments{
\item{seurat.obj}{the seurat.obj to add group}

\item{group.list}{a list including the information of the sampels}
}
\value{
seurat.obj
}
\description{
To split the cells to 'group' by 'samples'
}
\examples{
\dontrun{
group.list <- list('control' = c("A_L_001","A_L_002","A_L_003","A_L_004"),
                   'expr' = c("A_L_009","A_L_010","A_L_011","A_L_012","A_L_013"))
seurat.obj <- Get_group(seurat.obj,group.list)
}
}
