% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Scatterplot.R
\name{Get_Scatterplot1}
\alias{Get_Scatterplot1}
\title{to show genes like FACS with colors}
\usage{
Get_Scatterplot1(
  seurat_object,
  genes = c("x", "y"),
  filename = NULL,
  color_by = NULL
)
}
\arguments{
\item{seurat_object}{seurat.obj}

\item{genes}{the genes to show}

\item{filename}{the name of the generated file}

\item{color_by}{to color the cluster, default 'seurat cluster'}
}
\description{
to show genes like FACS with colors
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
Get_Scatterplot1(pbmc_1k,genes=c('CD8A','CD4'))
}
}
