% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Scatterplot.R
\name{Get_Scatterplot2}
\alias{Get_Scatterplot2}
\title{To show the corralations between two genes}
\usage{
Get_Scatterplot2(
  mat.object = NULL,
  seurat_object = NULL,
  log = F,
  genes = c("x", "y"),
  filename = NULL,
  type = NULL
)
}
\arguments{
\item{mat.object}{expression matrix as genes in row, samples or cells in column}

\item{seurat_object}{it also support a seurat_object instead of expression matrix}

\item{log}{the mat.object whether had log}

\item{genes}{the genes to show}

\item{filename}{the name of the generated file}

\item{type}{Smoothing method (function) to use, accepts either NULL(default) or a character vector, e.g. "lm", "glm", "gam", "loess" or a function}
}
\description{
To show the corralations between two genes
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
Get_Scatterplot1(pbmc_1k,genes=c('CD8A','CD4'))
}
}
