% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_SingleR.R
\name{Get_singleR}
\alias{Get_singleR}
\title{Title}
\usage{
Get_singleR(seurat.obj, species = c("m", "h"), types = c("main", "fine"))
}
\arguments{
\item{seurat.obj}{seurat.obj}

\item{species}{the species of the seurat.obj}

\item{types}{the types of annotation, 'main' or 'fine'}
}
\value{
the seurat,obj including the celltype annotation
}
\description{
Title
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
seurat.obj <- Get_singleR( pbmc_1K, species = 'h',types = 'main')
}
}
