% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_GO_barplot.R
\name{GO_diff_barplot}
\alias{GO_diff_barplot}
\title{This function is used to plot different GO terms between two groups}
\usage{
GO_diff_barplot(
  up_path = NULL,
  down_path = NULL,
  topn = 10,
  width = 8,
  height = 6,
  text.size = 1.5,
  filename = "barplot"
)
}
\arguments{
\item{up_path}{file(.csv) of up pathways in GO analyse}

\item{down_path}{file(.csv) of down pathways in GO analyse}

\item{topn}{number of top pathways to show}

\item{width}{the width of the figure}

\item{height}{the height of the figure}

\item{text.size}{the size of the text}

\item{filename}{the name of the generated file}
}
\description{
This function is used to plot different GO terms between two groups
}
\details{
barplot part which exhibit the topn different expressed genesets in two group
}
\examples{
\dontrun{
up_path="../GO_up_GO_UP_result.csv"
down_path="../GO_down_GO_UP_result.csv"
GO_diff_barplot(up_path = up_path,down_path = down_path,topn = 10,
}

}
