% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSVA.R
\name{GSVA_difference}
\alias{GSVA_difference}
\title{This function is used to perform GSVA analysis between two clusters}
\usage{
GSVA_difference(
  em = NULL,
  seurat_object = NULL,
  gspath = NULL,
  species = "mouse",
  group = NULL,
  adjPvalueCutoff = 0.001,
  logFCcutoff = 0.1,
  text.size = 2,
  topn = 20,
  min.sz = 2,
  width = 6,
  height = 5,
  filename = NULL,
  ylimit = 10
)
}
\arguments{
\item{em}{expression matrix as genes in row, samples or cells in column}

\item{seurat_object}{it also support a seurat_object instead of expression matrix}

\item{gspath}{the correct path of your geneset file(.gmt)}

\item{species}{"mouse"(default),'human'}

\item{group}{a vector of two group,if em is a expression data,it need supply}

\item{adjPvalueCutoff}{cutof the P.adjust}

\item{logFCcutoff}{cutoff the logFC}

\item{text.size}{the size of the text}

\item{topn}{number of the top pathways}

\item{min.sz}{minimum size of the resulting gene sets}

\item{width}{the width of the figure}

\item{height}{the height of the figure}

\item{filename}{the name of the generated file}

\item{ylimit}{limit of the y aes}
}
\value{
seurat.obj with the score of the pathways
}
\description{
This function is used to perform GSVA analysis between two clusters
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
GSVA_pheatmap( pbmc_1k,species="mouse",group=c('1','2'))
}
}
