% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSVA.R
\name{GSVA_find_gs_markers}
\alias{GSVA_find_gs_markers}
\title{To find geneset markers of each cluster}
\usage{
GSVA_find_gs_markers(
  em = NULL,
  seurat_object = NULL,
  gspath = NULL,
  species = "mouse",
  adjPvalueCutoff = 0.001,
  filename = NULL,
  logFCcutoff = 0.1,
  text.size = 2,
  topn = 20,
  min.sz = 2,
  parallel.sz = 12
)
}
\arguments{
\item{em}{expression matrix as genes in row, samples or cells in column}

\item{seurat_object}{it also support a seurat_object instead of expression matrix}

\item{gspath}{the correct path of your geneset file(.gmt)}

\item{species}{species of the seurat.obj or matrix}

\item{adjPvalueCutoff}{cutof the P.adjust}

\item{filename}{the name of the generated file}

\item{logFCcutoff}{cutof the logFC}

\item{text.size}{the size of the text}

\item{topn}{number of the top pathways}

\item{min.sz}{minimum size of the resulting gene sets}

\item{parallel.sz}{number of threads of execution to use when doing the calculations in parallel.}
}
\value{
seurat.obj with the marker_gs
}
\description{
To find geneset markers of each cluster
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
GSVA_pheatmap( seurat_object = pbmc_1k,species="mouse")
}
}
