% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSVA.R
\name{GSVA_pheatmap}
\alias{GSVA_pheatmap}
\title{This function is used to perform GSVA analysis}
\usage{
GSVA_pheatmap(
  em = NULL,
  seurat_object = NULL,
  gspath = NULL,
  species = "human",
  text.size = 2,
  topn = 20,
  min.sz = 2,
  filename = "geneset",
  clusters = NULL,
  pathname = NULL,
  width = 6,
  height = 5
)
}
\arguments{
\item{em}{expression matrix as genes in row, samples or cells in column}

\item{seurat_object}{it also support a seurat_object instead of expression matrix}

\item{gspath}{the correct path of your geneset file(.gmt)}

\item{species}{species of the seurat.obj or matrix}

\item{text.size}{the size of the text}

\item{topn}{number of the top pathways}

\item{min.sz}{minimum size of the resulting gene sets}

\item{filename}{the name of the generated file}

\item{clusters}{to choose the clusters you interested}

\item{pathname}{to choose the pathways you interested}

\item{width}{the width of the figure}

\item{height}{the heigh of the figure}
}
\description{
This function is used to perform GSVA analysis
}
\details{
This function is used to perform GSVA analysis, including pheatmap part which exhibit
the GSVA score of selected genesets in every samples or cells barplot part which exhibit the
top20 different expressed genesets in two group
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
GSVA_pheatmap(pbmc_1k,species="mouse")
}
}
