% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSVA.R
\name{GSVA_score_insert}
\alias{GSVA_score_insert}
\title{Insert pathway score into the seurat.obj}
\usage{
GSVA_score_insert(
  seurat_object = NULL,
  gspath = NULL,
  species = "mouse",
  min.sz = 2,
  parallel.sz = 12
)
}
\arguments{
\item{seurat_object}{it also support a seurat_object instead of expression matrix}

\item{gspath}{the correct path of your geneset file(.gmt)}

\item{species}{"mouse"(default),'human'}

\item{min.sz}{minimum size of the resulting gene sets}

\item{parallel.sz}{number of threads of execution to use when doing the calculations in parallel}
}
\value{
seurat.obj with the score of the pathways
}
\description{
Insert pathway score into the seurat.obj
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
GSVA_pheatmap(pbmc_1k,species="mouse")
}
}
