% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{HTO_get_singlet}
\alias{HTO_get_singlet}
\title{Citeseq divide HTOtag}
\usage{
HTO_get_singlet(
  path,
  filename = NULL,
  positive.quantile = 0.99,
  width = 7,
  height = 6
)
}
\arguments{
\item{path}{pathways cellranger exported data}

\item{filename}{the name of the generated file}

\item{positive.quantile}{The quantile of inferred 'negative' distribution for each hashtag - over which the cell is considered 'positive'. Default is 0.99}

\item{width}{the width of the figure}

\item{height}{the height of the figure}
}
\value{
seurat.obj including
}
\description{
Citeseq divide HTOtag
}
