% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorful}
\alias{colorful}
\alias{common_cluster}
\alias{T_NK_cluster}
\title{List with colors}
\format{
A list with 9 characters.
\describe{
\item{\code{colors}}{character COLUMN_DESCRIPTION}
\item{\code{color_chocolate}}{character COLUMN_DESCRIPTION}
\item{\code{color_blue}}{character COLUMN_DESCRIPTION}
\item{\code{color_pink}}{character COLUMN_DESCRIPTION}
\item{\code{color_red}}{character COLUMN_DESCRIPTION}
\item{\code{color_green}}{character COLUMN_DESCRIPTION}
\item{\code{color_yellow}}{character COLUMN_DESCRIPTION}
\item{\code{colors16}}{character COLUMN_DESCRIPTION}
\item{\code{colors_new}}{character COLUMN_DESCRIPTION}
}

A list with 16 characters.
\describe{
\item{\code{T_NK}}{character COLUMN_DESCRIPTION}
\item{\code{plasma}}{character COLUMN_DESCRIPTION}
\item{\code{Neutrophil}}{character COLUMN_DESCRIPTION}
\item{\code{cycling_T}}{character COLUMN_DESCRIPTION}
\item{\code{cycling_B}}{character COLUMN_DESCRIPTION}
\item{\code{Tumor_cell}}{character COLUMN_DESCRIPTION}
\item{\code{Fibroblast}}{character COLUMN_DESCRIPTION}
\item{\code{cycling_cell}}{character COLUMN_DESCRIPTION}
\item{\code{B_cell}}{character COLUMN_DESCRIPTION}
\item{\code{Mast_cell}}{character COLUMN_DESCRIPTION}
\item{\code{Myloid}}{character COLUMN_DESCRIPTION}
\item{\code{Erythrocyte}}{character COLUMN_DESCRIPTION}
\item{\code{Endothelial}}{character COLUMN_DESCRIPTION}
\item{\code{EVT}}{character COLUMN_DESCRIPTION}
\item{\code{ESC}}{character COLUMN_DESCRIPTION}
\item{\code{pDC}}{character COLUMN_DESCRIPTION}
}

A list with 7 characters.
\describe{
\item{\code{MAIT}}{character COLUMN_DESCRIPTION}
\item{\code{CD4}}{character COLUMN_DESCRIPTION}
\item{\code{CD8}}{character COLUMN_DESCRIPTION}
\item{\code{NK}}{character COLUMN_DESCRIPTION}
\item{\code{iNKT}}{character COLUMN_DESCRIPTION}
\item{\code{Treg}}{character COLUMN_DESCRIPTION}
\item{\code{gdT}}{character COLUMN_DESCRIPTION}
}
}
\usage{
colorful

common_cluster

T_NK_cluster
}
\description{
List with colors

List with common_cluster

List with common_cluster
}
\details{
DETAILS

DETAILS

DETAILS
}
\section{Functions}{
\itemize{
\item \code{colorful}: for many kinds of colors

\item \code{common_cluster}: for 16 kinds of celltypes

\item \code{T_NK_cluster}: for 7 kinds of T_NK cluster
}}

\keyword{datasets}
