% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfeatureplot.R
\name{myfeatureplot}
\alias{myfeatureplot}
\title{Featureplot for genes and split by samples, group and so on.}
\usage{
myfeatureplot(seurat.obj, genenames, filename = NULL, spilt_by = NULL)
}
\arguments{
\item{seurat.obj}{seurat.obj to draw}

\item{genenames}{genenames to draw a featureplot}

\item{filename}{the name of the generated file}

\item{spilt_by}{whether choose to spilt.by in meta.data}
}
\description{
Featureplot for genes and split by samples, group and so on.
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
genenames <- c("MS4A1", "CD79B", "CD79A", "HLA-DRA", "TCL1A", "HLA-DQB1", "HVCN1")
myfeatureplot(pbmc_1k, genenames)
myfeatureplot(pbmc_1k, genenames, spilt_by = 'group')
}
}
