% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_BASIC.R
\name{myfindmarkers}
\alias{myfindmarkers}
\title{To get the top DE gene of cluster}
\usage{
myfindmarkers(
  seurat_object,
  filename = NULL,
  thresh.use = 0.25,
  species = "human",
  all.markers = NULL,
  colors = NULL
)
}
\arguments{
\item{seurat_object}{seurat.obj}

\item{filename}{the name of the generated file}

\item{thresh.use}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells}

\item{species}{the species of the seurat.obj}

\item{all.markers}{the result of the FindAllMarkers()}

\item{colors}{the colors of clusters}
}
\value{
a list including two characters,top5 and top20
}
\description{
To get the top DE gene of cluster
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
top_genes <- myfindmarkers(pbmc_1k)
}
}
