% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGO_KEGG.R
\name{myGO_KEGG}
\alias{myGO_KEGG}
\title{GO and KEGG analyse for the DEG}
\usage{
myGO_KEGG(
  species = "m",
  up_gene,
  down_gene,
  png.width = 1000,
  png.height = 500,
  pdf.width = 700,
  pdf.height = 5,
  show = 20,
  filename = NULL
)
}
\arguments{
\item{species}{species of genes}

\item{up_gene}{character of up genes}

\item{down_gene}{character of down genes}

\item{png.width}{width of the figure in png}

\item{png.height}{height of the figure in png}

\item{pdf.width}{width of the figure in pdf}

\item{pdf.height}{height of the figure in pdf}

\item{show}{number of the pathway to show}

\item{filename}{the name of the generated file}
}
\description{
GO and KEGG analyse for the DEG
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLAB')
DEG_matrix <- read.csv(input.file, header = TRUE, row.names = 1)
up_gene <- rownames(subset(DEG_matrix,avg_logFC > 0))
down_gene <- rownames(subset(DEG_matrix,avg_logFC < 0))
DEG <- myGO_KEGG(species="h",up_gene,down_gene)
}
}
