% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{myreplace_ident}
\alias{myreplace_ident}
\title{This function is used to replace the cluster idents in target}
\usage{
myreplace_ident(target_object, replace_object)
}
\arguments{
\item{target_object}{the target seurat.obj}

\item{replace_object}{the replace seurat.obj, including the new annotation}
}
\value{
the seurat.obj had new annotation
}
\description{
This function is used to replace the cluster idents in target
}
\examples{
\dontrun{
target_object = myreplace_ident(target_object,replace_object)
}

}
