% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myvolcano.R
\name{myvolcano}
\alias{myvolcano}
\title{Volcano for DE genes.}
\usage{
myvolcano(df, gene.plot = 10, logfc.cutoff = 0.5, p.cutoff = 0.01)
}
\arguments{
\item{df}{DE gene matrix, including cols named avg_logFC and p_val_adj}

\item{gene.plot}{number of tags for genes}

\item{logfc.cutoff}{cutoff of logFC}

\item{p.cutoff}{cutoff of p value}
}
\value{
DE genes
}
\description{
Volcano for DE genes.
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLABDATA')
DEG_matrix <- read.csv(input.file, header = TRUE, row.names = 1)
DEG <- myvolcano(df = DEG_matrix, gene.plot = 10, logfc.cutoff = 0.5, p.cutoff = 0.01)
}
}
