% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myviolin_plot.R
\name{plot_violin_manual}
\alias{plot_violin_manual}
\title{plot the stacked violinplot of single cell RNAseq data}
\usage{
plot_violin_manual(
  gene,
  seurat_object,
  cluster = NULL,
  axis = FALSE,
  legend_position = "none",
  limits.max = 7,
  gene_name = FALSE,
  width = 3.6,
  height = 2,
  flip = F
)
}
\arguments{
\item{gene}{a vector of the genes name to plot}

\item{seurat_object}{the seurat object of your data}

\item{cluster}{choose to cluster you interested}

\item{axis}{logical,set the hidden axes}

\item{legend_position}{position of legend}

\item{limits.max}{numeric,the max legend value to regulate the color levels}

\item{gene_name}{logical, set gene titile bold}

\item{width}{the width of the figure}

\item{height}{the height of the figure}

\item{flip}{logical, set the direction of coordinate axis}
}
\description{
plot the stacked violinplot of single cell RNAseq data
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_1k.RDS',package = 'LIANLABDATA')
pbmc_1k <- readRDS(input.file)
}
}
