% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{See_mRNA}
\alias{See_mRNA}
\title{To see the distribution of the number of genes}
\usage{
See_mRNA(seurat.obj)
}
\arguments{
\item{seurat.obj}{the seurat.obj  to cut}
}
\description{
To see the distribution of the number of genes
}
\examples{
\dontrun{
input.file <- system.file('extdata','pbmc_small.RDS',package = 'LIANLABDATA')
load(input.file)
See_mRNA(pbmc_small)
}
}
