% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{test_package}
\alias{test_package}
\title{To test the package need}
\usage{
test_package(package_list)
}
\arguments{
\item{package_list}{the character of the test package}
}
\description{
To test the package need
}
\examples{
\dontrun{
test_package(c('Seurat','DOSE'))
}
}
