% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{top_m}
\alias{top_m}
\title{Get DE genes in ave_logFC}
\usage{
top_m(markers, m)
}
\arguments{
\item{markers}{the result of FindAllMarkers(), including the col named "cluster"}

\item{m}{the number of the genes in top}
}
\value{
a data.frame has the genes in top you choose
}
\description{
Get DE genes in ave_logFC
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLABDATA')
markers <- read.csv(input.file,header = T, row.names = 1)
markers_top <- top_m( markers, m = 20)
}
}
