% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Mixture.R
\name{top_up_down_m}
\alias{top_up_down_m}
\title{Get DE genes in ave_logFC}
\usage{
top_up_down_m(marker, m)
}
\arguments{
\item{marker}{the result of DE genes of a cluster or the result between the two clusters}

\item{m}{the number of the genes in top and down}
}
\value{
a data.frame has the genes in top/down you choose
}
\description{
Get DE genes in ave_logFC
}
\examples{
\dontrun{
input.file <- system.file('extdata','DEG.csv',package = 'LIANLABDATA')
markers <- read.csv(input.file,header = T, row.names = 1)
markers <- subset(markers,subset = cluster == 'Naive CD4 T')
markers_top <- top_m( markers, m = 20)
}
}
